Use uint256 for round outcome flags
Widen RoundStats.flags and SAME_BID_EARNED_ENDURANCE_CHAMPION_AND_CHRONO_WARRIOR_TITLES from uint8 to uint256. Keep the same-bid flag value at 1 << 2 and remove the obsolete uint8 cast from its declaration. The generated roundStats getter now exposes flags as uint256.

Update the assert-enabled round-finalization check to compare flags directly with zero instead of uint8(0). Remove the resolved comments about revisiting the flag types and packing the flags field.

+++ contracts/production/interfaces/ICosmicSignatureGameStorage.sol
@@ -90,9 +90,7 @@ interface ICosmicSignatureGameStorage {
 
 		/// @notice In V3+, bidding round outcome flags.
 		/// Comment-202610069 relates.
-		/// @dev todo-0 Revisit this variable type.
-		/// todo-0 Consider combining this variable and/or other variables with other variables in the same storage slot.
-		uint8 flags;
+		uint256 flags;
 	}
 
 	/// @notice Details about a bidder.
